(PC-6462) Rename adresse folder to place
