+++ src/features/search/components/locationChoice.utils.tsx
@@ -2,9 +2,9 @@ import { t } from '@lingui/macro'
 import React from 'react'
 
 import { useSearch } from 'features/search/pages/SearchWrapper'
-import { buildPlaceLabel } from 'libs/adresse/buildPlaceLabel'
 import { LocationType } from 'libs/algolia'
 import { _ } from 'libs/i18n'
+import { buildPlaceLabel } from 'libs/place'
 import { AroundMe } from 'ui/svg/icons/AroundMe'
 import { BicolorLocationPointer as Place } from 'ui/svg/icons/BicolorLocationPointer'
 import { Everywhere } from 'ui/svg/icons/Everywhere'
