+++ src/features/search/pages/LocationPicker.tsx
@@ -6,11 +6,9 @@ import { FlatList, TouchableOpacity } from 'react-native'
 import styled from 'styled-components/native'
 
 import { useSearch } from 'features/search/pages/SearchWrapper'
-import { REGEX_STARTING_WITH_NUMBERS } from 'libs/adresse/buildPlaceLabel'
-import { fetchPlaces } from 'libs/adresse/fetchPlaces'
-import { SuggestedPlace } from 'libs/adresse/types'
 import { LocationType } from 'libs/algolia'
 import { _ } from 'libs/i18n'
+import { fetchPlaces, REGEX_STARTING_WITH_NUMBERS, SuggestedPlace } from 'libs/place'
 import { PageHeader } from 'ui/components/headers/PageHeader'
 import { SearchInput } from 'ui/components/inputs/SearchInput'
 import { Invalidate } from 'ui/svg/icons/Invalidate'
